make WallMST 48-bit

+++ src/serde_impl.rs
@@ -1,7 +1,7 @@
 use serde::{de, ser};
 
 #[derive(Serialize, Deserialize)]
-struct Timestamp<T>(T, u32);
+struct Timestamp<T>(T, u16);
 
 impl<T: ser::Serialize + Copy> ser::Serialize for crate::Timestamp<T> {
     fn serialize<S: ser::Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
